refactor(map-source): rename ports.ts and ports.spatial.ts (#1197)
- ports.ts → types.ts (P2: types-only aggregation)
- ports.spatial.ts → spatialTypes.ts (P2: types-only)

Rollback: git revert <hash> --no-edit
