+++ packages/map-source/src/adapters/FeatureCollectionGridIndex.ts
@@ -1,4 +1,4 @@
-import type { BBox, FeatureCollection, MapSourcePort, TileCoord } from '~/ports';
+import type { BBox, FeatureCollection, MapSourcePort, TileCoord } from '~/types';
 
 interface Cell {
   feats: any[];
