+++ packages/map-source/src/spatialTypes.ts
@@ -1,4 +1,4 @@
-import type { FeatureCollection, MapSourcePort } from './ports.js';
+import type { FeatureCollection, MapSourcePort } from './types.js';
 
 export interface SpatialIndexPort extends MapSourcePort {
   build(fc: FeatureCollection): Promise<void> | void;
